Created a file to scrape http://www.hltv.org/events/upcoming/ to see the upcoming contests and add them to our database if there is enough information
I also scrape additional info about contests that I could store in the
database if people wanted. Things like hltvID, LAN or online, location,
prizePot, and number of teams competing. Just lmk if we want to add any
of this stuff!

Currently only adding contests if we know who 4/5 of the players
competing are. Might want to change this requirement to either be lower
or require all players?

Two possible issues:
1. The dates are in a different format, which I can fix if people want
me to, it’d just be a mild pain :(.
2. The contest info is also stored in a different order in the database
than it was previously, which would be a decently sized pain to fix,
though it’s probably doable if necessary?